MLBaseEditor: create component with padding, margin, background. MLTextEditor: updated, deleted padding, margin, background
