+++ src/ui/pages/main/multilink/editor/content/MLTextEditor/MLTextEditor.module.scss
@@ -1,7 +1,22 @@
-.flex__row3 {
+.flex_row {
   margin-top: 5px;
   display: flex;
   text-transform: capitalize;
+}
+
+.flex_row2 {
+  display: flex;
+  flex-direction: row;
+  margin-top: 15px;
+  justify-content: space-around;
+}
+
+.font_shadow_select {
+  display: flex;
+  flex-direction: column;
+  margin-top: 15px;
+  width: 150px;
+  justify-content: center;
 }
 
 .select {
@@ -17,29 +32,5 @@
   height: 30px;
   margin-top: 5px;
   border-radius: 7px;
-}
-
-.circle {
-  width: 30px;
-  height: 30px;
-  border: 1px solid black;
-  display: inline-block;
-  margin-left: 5px;
-  padding: 2px;
-  border-radius: 50%;
-}
-
-.circleGradient {
-  width: 30px;
-  height: 30px;
-  border: 1px solid black;
-  display: inline-block;
-  margin-left: 5px;
-  padding: 2px;
-  border-radius: 50%;
-  background:
-          linear-gradient(217deg, rgba(255,0,0,.8), rgba(255,0,0,0) 70.71%),
-          linear-gradient(127deg, rgba(0,255,0,.8), rgba(0,255,0,0) 70.71%),
-          linear-gradient(336deg, rgba(0,0,255,.8), rgba(0,0,255,0) 70.71%);
 }
