+++ src/ui/styles/_ui.scss
@@ -1,5 +1,3 @@
-
-
 h1 {
   @include stretch-it(font-size, 20.8, 52);
   font-weight: 700;
@@ -279,4 +277,36 @@ h4 {
   height: 100%;
   width: 100%;
   z-index: 100;
+}
+
+.circle {
+  width: 30px;
+  height: 30px;
+  border: 1px solid black;
+  display: inline-block;
+  margin-left: 5px;
+  padding: 2px;
+  border-radius: 50%;
+}
+
+.circleGradient {
+  width: 30px;
+  height: 30px;
+  border: 1px solid black;
+  display: inline-block;
+  margin-left: 5px;
+  padding: 2px;
+  border-radius: 50%;
+  background:
+          linear-gradient(217deg, rgba(255,0,0,.8), rgba(255,0,0,0) 70.71%),
+          linear-gradient(127deg, rgba(0,255,0,.8), rgba(0,255,0,0) 70.71%),
+          linear-gradient(336deg, rgba(0,0,255,.8), rgba(0,0,255,0) 70.71%);
+}
+
+.padding_margin {
+    display: flex;
+    flex-direction: row;
+    margin: 10px 0 10px 0;
+    justify-content: space-around;
+    text-transform: capitalize;
 }
